Expose an /api/v1/extensions endpoint.
We don't have any extensions listed or link this in the "endpoints" field on
actors yet. The endpoint also doesn't currently provide a way to expose any
additional metadata about an extension beyond its URL (in part because the
SWICG proposal currently doesn't include that).

We can add extensions to the endpoint's response by calling
Mastodon::Extension::register with the extension URL.